유저/메뉴/댓글 간 종속성 때문에 삭제 오류가 생김. 수정 중에 있음.

+++ src/main/java/org/TUK/gotMenu/controller/MenuController.java
@@ -40,6 +40,8 @@ public class MenuController {
 
     @GetMapping("/create")
     public String create(MenuForm menuForm){
+
+        if(securityService.getUserNo() == null) return "redirect:/menu/list";
        return "/menu/menuForm";
     }
 
@@ -55,6 +57,9 @@ public class MenuController {
     @GetMapping("/{menuNo}/modify")
     public String modify(MenuForm menuForm, @PathVariable("menuNo") Integer menuNo){
         Menu menu = this.menuService.getMenu(menuNo);
+        if(!securityService.isSameUser(menu.getWriter().getUserNo()))
+            return "redirect:/menu/list";
+
         menuForm.setMenuDescription(menu.getMenuDescription());
         menuForm.setMenuDetail(menu.getMenuDetail());
         menuForm.setMenuComposition(menu.getMenuComposition());
@@ -77,6 +82,9 @@ public class MenuController {
     @GetMapping("/{menuNo}/delete")
     public String delete(@PathVariable("menuNo") Integer menuNo){
         Menu menu = this.menuService.getMenu(menuNo);
+        if(!securityService.isSameUser(menu.getWriter().getUserNo()))
+            return "redirect:/menu/list";
+
         this.menuService.delete(menu);
         return "redirect:/menu/list";
     }
@@ -91,6 +99,7 @@ public class MenuController {
     @ResponseBody
     private Integer test()
     {
+        if(securityService.getUserNo() == null) return null;
         for (int i = 0 ; i < 1000 ; i ++){
             this.menuService.create("미역국, 밥", "칼로리는 500kcal", "맛있는 미역국에 밥을 말아먹으면 꿀맛!", i, "가정식");
         }
